feat: #538 add common module for junit5
