+++ extension/junit5-platform-8/pom.xml
@@ -13,6 +13,10 @@
     <dependencies>
         <dependency>
             <groupId>${project.groupId}</groupId>
+            <artifactId>camunda-process-test-coverage-junit5-common</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
             <artifactId>camunda-process-test-coverage-engine-platform-8</artifactId>
         </dependency>
         <dependency>
